pyproject.toml
[tool.poetry]
name = "swipe"
version = "0.0.1"
description = "---SWIPE_UA---"
authors = ["i am <example@example.com>"]
package-mode = false

# отключаем установку проекта как пакета
packages = []

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = "^0.128.5"
uvicorn = ">=0.24.0,<0.25.0"
sqlalchemy = ">=2.0.20,<3.0.0"
asyncpg = ">=0.27.0"
alembic = ">=1.17.0"
httpx = ">=0.27.0,<0.28.0"
python-dotenv = ">=1.0.0,<2.0.0"
dishka = "^1.7.2"
pydantic-settings = "^2.12.0"
redis = "<5"
aiosqlite = "^0.21.0"
psycopg2-binary = "^2.9.11"
pydantic = "^2.12.5"
email-validator = "^2.3.0"
pyjwt = "^2.10.1"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
passlib = "1.7.4"
bcrypt = "4.0.1"
python-multipart = "^0.0.20"
celery = "5.3.6"
aiosmtplib = "5.0.0"
itsdangerous = "^2.2.0"
python-magic = "^0.4.27"
pytest = "^9.0.2"
fastapi-pagination = "^0.15.10"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.10.0"
httpx = "^0.27.0"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "users: Тести модифікації користувачів (CRUD)",
    "user: Тести профілю користувача",
    "auth: Тести авторизації та JWT токенів",
    "filter: тести для фільтрів оголошень",
]

testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
